feat: Implement product data source and repository
- Added ProductDataSource to fetch all products from the server.
- Created ProductRepository to handle data retrieval and error management.
- Updated ProductResponse and ProductModel to parse JSON data correctly.
- Introduced ProductVariantModel with necessary fields and JSON parsing.
- Developed CategoryController to manage category state and fetching.
- Created ProductControllers to manage product state and fetching.
- Updated home screen to temporarily disable PopularSection.
- Refactored product list screen to comment out unused code.
- Added CategoryState and ProductState for state management.
- Updated CategorySection to use the correct controller import.
- Commented out PopularSection widget code for future use.
- Enhanced ProductCard to display product and variant details.
- Implemented RecommendedSection to fetch and display recommended products.
- Updated favorite screen with sample product data.
- Adjusted product card in wishlist to use correct data types.
- Added utility functions for color and currency formatting.
- Updated pubspec.lock to reflect dependency changes.

+++ lib/src/features/wishlist/presentation/widgets/product_card.dart
@@ -5,9 +5,9 @@ class Product {
   final String name;
   final String color;
   final String storage;
-  final double price;
+  final int price;
   final String imageUrl;
-  final double rating;
+  final int rating;
   final int reviews;
 
   Product({
